✨ Finish new input code
